fixup! add shared_page_handler to Core::System; change epoch to millisecond

+++ src/citra_qt/configuration/config.cpp
@@ -410,7 +410,8 @@ void Config::SaveValues() {
     WriteSetting("region_value", Settings::values.region_value, Settings::REGION_VALUE_AUTO_SELECT);
     WriteSetting("init_clock", static_cast<u32>(Settings::values.init_clock),
                  static_cast<u32>(Settings::InitClock::SystemTime));
-    WriteSetting("init_time", static_cast<u64>(Settings::values.init_time), 946681277ULL);
+    WriteSetting("init_time", static_cast<unsigned long long>(Settings::values.init_time),
+                 946681277ULL);
     qt_config->endGroup();
 
     qt_config->beginGroup("Miscellaneous");
